+++ src/lib.rs
@@ -18,8 +18,6 @@ pub mod utils;
 #[cfg(feature = "wayland_frontend")]
 pub mod wayland;
 
-pub mod signaling;
-
 #[cfg(feature = "xwayland")]
 pub mod xwayland;
